Tạm thời dừng việc thêm nội dung mới để chuẩn bị cho Review

+++ backend/src/main/java/com/codegym/backend/dto/ItemResponse.java
@@ -1,7 +1,6 @@
 package com.codegym.backend.dto;
 
 import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -23,6 +22,5 @@ public class ItemResponse {
     private String description;
     private String imageUrl;
     private Boolean isAvailable;
-    private Integer totalOrderCount;
-
+    private Long totalOrderCount;
 }
